ai-images-dev/Dev/Mongo_Swift_Insert_Dataset/requirements.txt
flask>=1.1,<1.2
markupsafe==2.0.1
python-swiftclient>=3.11,<3.12
pymongo>=3.11,<3.12
influxdb_client==1.30.0
pandas
pynacl
Pillow
pyOpenSSL
paramiko
uwsgi
flask_swagger_ui
Werkzeug==0.16.1
apispec
apispec_webframeworks
marshmallow
keystoneauth1==4.5.0
python-keystoneclient==4.4.0
python-swiftclient
pyodbc
natsort
python-dotenv